Refactor import update
